Separate out VaultRuntimeSubstate
